Handled edge cases

+++ index.js
@@ -3,9 +3,11 @@ var sqlite3 = require('sqlite3').verbose()
 var program = require('commander')
 var Jusibe = require('jusibe')
 var rl = require('readline')
+var fs = require('fs')
 
 
 var db = new sqlite3.Database('contacts_db.sqlite')
+
 //first goal create a database to store the contacts
 program
   .command('add <name> <phone_number>')
@@ -36,32 +38,48 @@ program
   .command('search <name>')
   .description('search for <name>')
   .action(function(name, command){
-    //do a check to see if db exists before searching
-    //search for the given name
 
-    db.all("SELECT contact_id, contact_name, contact_number FROM contacts WHERE contact_name LIKE " + "'" + "%" + name + "%';", function(err, row){
-      //handle case where there is more than one contact with a given name
-      if(row.length > 1){
-        var rl = require('readline')
-        var read = rl.createInterface({
-          input: process.stdin,
-          output: process.stdout
-        });
+    db.serialize(function(){
+      db.get("SELECT * FROM sqlite_master WHERE type='table' AND name = 'contacts';", function(err, row){
 
-        for(i = 0; i < row.length; i++){
-          console.log("[" + i + "]", row[i].contact_name.replace(name, ''), row[i].contact_number)
+        //check if the database table exists before searching
+        if(row === undefined){
+          console.log("Database table for contacts does not exist! Please add contacts before attempting to search")
         }
 
-        read.question("\nWhich " + name + "? (Enter the corresponding number to indicate the contact) ", function(answer){
-          read.close()
-          answer = parseInt(answer); //convert the answer from string to integer
-          console.log("You chose " + row[answer].contact_name, row[answer].contact_number)
-        });
-      }
+        else{
+          db.all("SELECT contact_id, contact_name, contact_number FROM contacts WHERE contact_name LIKE " + "'" + "%" + name + "%';", function(err, row){
 
-      else{
-        console.log("[" + row[0].contact_name + "]", row[0].contact_number);
-      }
+            if(row.length > 1){
+              var read = rl.createInterface({
+                input: process.stdin,
+                output: process.stdout
+              });
+
+              //loop through array and print members of the array
+              for(i = 0; i < row.length; i++){
+                console.log("[" + i + "]", row[i].contact_name.replace(name, ''), row[i].contact_number);
+              }
+
+              read.question("\nWhich " + name + "? (Enter the corresponding number to indicate the contact) ", function(answer){
+                read.close()
+                answer = parseInt(answer); //convert the answer from string to integer
+                console.log("You chose " + row[answer].contact_name, row[answer].contact_number)
+              });
+            }
+
+            else if(row.length === 1){
+              console.log("[" + row[0].contact_name + "]", row[0].contact_number);
+            }
+
+            else{
+              console.log("Contact not found!! Please check your search query")
+            }
+
+          });
+        }
+
+      });
 
     });
 
@@ -72,9 +90,20 @@ program
   .command('view')
   .description('View all stored contacts')
   .action(function(command){
-    //sql statement to view all contacts
-    db.all("SELECT * FROM contacts ORDER BY contact_name;", function(err, row){
-      console.log(row)
+
+    db.serialize(function(){
+      db.get("SELECT * FROM sqlite_master WHERE type='table' AND name = 'contacts';", function(err, row){
+        //check if the database table exists
+        if(row === undefined){
+          console.log("No contacts to view because the database table does not exist")
+        }
+
+        else{
+          db.all("SELECT * FROM contacts ORDER BY contact_name;", function(err, row){
+            console.log(row)
+          });
+        }
+      });
     });
   });
 
@@ -84,35 +113,48 @@ program
   .command('del <name>')
   .description('Delete <name> from contacts')
   .action(function(name, command){
-    //first search the database for entries that match that name
-    db.all("SELECT contact_id, contact_name, contact_number FROM contacts WHERE contact_name LIKE " + "'" + "%" + name + "%';", function(err, row){
 
-      if(row.length > 1){
-        var rl = require('readline')
-        var read = rl.createInterface({
-          input: process.stdin,
-          output: process.stdout
-        });
-
-        //print the retrieved data
-        for(i = 0; i < row.length; i++){
-          console.log("[" + i + "]", row[i].contact_name.replace(name, ''), row[i].contact_number)
+    db.serialize(function(){
+      db.get("SELECT * FROM sqlite_master WHERE type='table' AND name = 'contacts';", function(err, row){
+        if(row === undefined){
+          console.log("Cannot delete contact because database table does not exist")
         }
 
-        read.question("\nWhich " + name + "? (Enter the corresponding number to indicate the contact you want to delete) ", function(answer){
-          read.close()
-          answer = parseInt(answer); //convert the answer from string to integer
-          db.run("DELETE FROM contacts WHERE contact_name = " + "'" + row[answer].contact_name + "'")
-          console.log("Contact successfully deleted!!")
-        });
-      }
+        else{
+          //first search the database for entries that match that name
+          db.all("SELECT contact_id, contact_name, contact_number FROM contacts WHERE contact_name LIKE " + "'" + "%" + name + "%';", function(err, row){
+            if(row.length > 1){
+              var read = rl.createInterface({
+                input: process.stdin,
+                output: process.stdout
+              });
 
-      else{
-        db.run("DELETE FROM contacts WHERE contact_name = " +  "'" + row[0].contact_name + "'")
-        console.log("Contact successfully deleted!!")
-      }
+              //print the retrieved data
+              for(i = 0; i < row.length; i++){
+                console.log("[" + i + "]", row[i].contact_name.replace(name, ''), row[i].contact_number)
+              }
 
-    });
+              read.question("\nWhich " + name + "? (Enter the corresponding number to indicate the contact you want to delete) ", function(answer){
+                read.close();
+                answer = parseInt(answer); //convert the answer from string to integer
+                db.run("DELETE FROM contacts WHERE contact_name = " + "'" + row[answer].contact_name + "'")
+                console.log("Contact successfully deleted!!")
+              });
+
+            }
+
+            else if(row.length === 1){
+              db.run("DELETE FROM contacts WHERE contact_name = " +  "'" + row[0].contact_name + "'")
+              console.log("Contact successfully deleted!!")
+            }
+
+            else{
+              console.log("Contact cannot be deleted because it does not exist in the database")
+            }
+          });
+        }
+      });
+    }); 
   });
 
 //next goal --> user should be able to send an SMS to another user in his contact list
@@ -121,77 +163,86 @@ program
   .option('-m, --short_message', 'Message you want to send')
   .description('Send <short_message> to <name>')
   .action(function(short_message, name, command){
-    db.all("SELECT contact_name, contact_number FROM contacts WHERE contact_name LIKE " + "'" + "%" + name + "%';", function(err, row){
-      //set values for jusibe's API
-      jusibe_pub_key = "2f1a1c3ab844aa292dd592e7a1abacc6";
-      jusibe_acc_token = "4352384c4191e982c08328308f50d09a";
-      var jusibe = new Jusibe(jusibe_pub_key, jusibe_acc_token)
 
-      //handle cases where more than one contact has the specified
-      if(row.length > 1){
-        //create interface for collecting input from the user
-        var read = rl.createInterface({
-          input: process.stdin,
-          output: process.stdout
-        });
-
-        //print the retrieved data
-        for(i = 0; i < row.length; i++){
-          console.log("[" + i + "]", row[i].contact_name.replace(name, ''), row[i].contact_number)
+    db.serialize(function(){
+      db.get("SELECT * FROM sqlite_master WHERE type='table' AND name = 'contacts';", function(err, row){
+        if(row === undefined){
+          console.log("Database table for contacts does not exist! Please add contacts before attempting to send a message");
         }
 
-        read.question("\nWhich " + name + "? (Enter the corresponding number to indicate the contact you want to send the SMS to) ", function(answer){
+        else{
+          db.all("SELECT contact_name, contact_number FROM contacts WHERE contact_name LIKE " + "'" + "%" + name + "%';", function(err, row){
+            //set values for jusibe's API
+            jusibe_pub_key = "2f1a1c3ab844aa292dd592e7a1abacc6";
+            jusibe_acc_token = "4352384c4191e982c08328308f50d09a";
 
-          read.close()
-          answer = parseInt(answer); //convert the answer from string to integer
-          
-          console.log("You chose " + row[answer].contact_name)
-          console.log("Sending message...")
+            var jusibe = new Jusibe(jusibe_pub_key, jusibe_acc_token)
 
-          //send message
-          var payload = {
-            to: row[answer].contact_number,
-            from: 'SMS_APP',
-            message: short_message
-          };
+            if(row.length > 1){
+              //create interface for collecting input from the user
+              var read = rl.createInterface({
+                input: process.stdin,
+                output: process.stdout
+              });
 
-          //send message confirming whether or not the message was sent
-         jusibe.sendSMS(payload, function(err, res){
+              for(i = 0; i < row.length; i++){
+                console.log("[" + i + "]", row[i].contact_name.replace(name, ''), row[i].contact_number)
+              }
 
-          if(res.statusCode === 200){
-            console.log(res.body);
+              read.question("\nWhich " + name + "? (Enter the corresponding number to indicate the contact you want to send the SMS to) ", function(answer){
+                read.close()
+                answer = parseInt(answer); //convert answer from string to integer
+
+                console.log("You chose " + row[answer].contact_name)
+                console.log("Sending message...")
+
+                //send message
+                var payload = {
+                  to: row[answer].contact_number,
+                  from: 'Contacto',
+                  message: short_message
+                }
+
+                //print message confirming whether or not message was sent
+                jusibe.sendSMS(payload, function(err, res){
+                  if(res.statusCode === 200){
+                    console.log(res.body);
+                  }
+
+                  else{
+                    console.log("Message not sent!!" + "\n Check the number and/or ensure you are connected to the internet \n");
+                  }
+                })
+              });
+            }
+            else if(row.length === 1){
+              console.log("Sending message...")
+
+              //send message
+              var payload = {
+              to: row[0].contact_number,
+              from: 'Contacto',
+              message:short_message
+            };
+
+            jusibe.sendSMS(payload, function(err, res){
+              if(res.statusCode === 200){
+                console.log(res.body)
+              }
+
+             else{
+              console.log("Message not sent!!" + "\n Check the number and/or ensure you are connected to the internet \n");
+             }
+           });
           }
-          else{
-            console.log("Message not sent!!" + "\n Use the error message below to debug \n");
-            console.log(err);
-          }
 
-         });
-
-        });
-      }
-
-      else{
-
-        var payload = {
-          to: row[0].contact_number,
-          from: 'SMS_APP',
-          message:short_message
-        };
-
-        jusibe.sendSMS(payload, function(err, res){
-          if(res.statusCode === 200){
-            console.log(res.body)
-          }
-
-          else{
-
-            console.log("Message not sent!!" + "\n Use the error message below to debug \n");
-            console.log(err);
-          }
-        });
-      }
-
+            else{
+              console.log("Message not sent because the contact name does not exist in the database")
+            }
+          });          
+        }
+      });
     });
   });
+
   program.parse(process.argv)
